backend/requirements.txt
fastapi
uvicorn
pymupdf
faiss-cpu
sentence-transformers
python-multipart
